Fixing `Makefile`.
